moving AGCCompress back to cli package to facilitate merge
